Trying to model the Cube, as well as add depth buffering
Added Cube.js, Cube.html, and main.js. I think I have the vertices and indices correct for the cube, but I'm not 100% sure. I'm confused on how to add colors to the vertex shader. In the vertex shader there is "gl_Position = aPosition" but I'm not sure what to put for colors. In the fragment shader I currently have: 

"
        vec4 frontColor = vec4(0.0, 1.0, 1.0, 1.0); // cyan
        vec4 backColor = vec4(1.0, 0.0, 1.0, 1.0); // purple
        fColor = gl_FrontFacing ? frontColor : backColor;
"

When I open up the Cone.html webpage I have a cyan square with a black background. I also added depth buffering in main.js.